Update Enable-CsMeetingRoom.md
This sentence is contradictory and causes confusion on support cases.
EmailAddress: The user's email address (as defined in Active Directory) is used as the SIP address.UserPrincipalName. The user's UPN is used as the SIP address.

This should only state that when using the EmailAddress parameter, the user's UPN is used as the SIP address.